// src/enums/menuItems.ts
export enum MenuItems {
  RESET = "RESET",
  PLAY = "PLAY",
  PAUSE = "PAUSE",
  PLAYBACK_SPEED = "PLAYBACK_SPEED",
  SHOW_PREFERENCES_MODAL = "SHOW_PREFERENCES",
  SHOW_INFORMATION_MODAL = "SHOW_INFORMATION",
  SHOW_MORE = "SHOW_MORE",
  SHOW_ENERGY_GRAPH_MODAL = "SHOW_ENERGY_GRAPH",
  SHOW_CONTROLS_MODAL = "SHOW_CONTROLS",
  SHOW_OPTIONS_MODAL = "SHOW_OPTIONS",
  SHOW_STOPWATCH = "SHOW_STOPWATCH",
}